Port MSBFS to unified Core SPI
- Remove int-based version in favour of long-based
- Flatten inheritance to only one class
- Rename away from Huge prefix

+++ contrib/algo/src/main/java/org/neo4j/graphalgo/impl/msbfs/BfsConsumer.java
@@ -22,5 +22,5 @@ package org.neo4j.graphalgo.impl.msbfs;
 @FunctionalInterface
 public interface BfsConsumer {
 
-    void accept(int nodeId, int depth, BfsSources sourceNodeIds);
+    void accept(long nodeId, int depth, BfsSources sourceNodeIds);
 }
